Fix odd query results
site:wikipedia.org was attached to the end of the query without a space
this would sometimes result in odd query results such as: flusite:wikipedia.org
returned something referring to EUR-Lex or whatever that is totally unrelated.
So now I added a space: flu site:wikipedia.org. That fixes the queries