Fix weather command
The issue was because the weather URL moved to another location, and was
sending a redirection notice to the AtomFeed. The AtomFeed for some
reason doesn't do redirection and would wrongly interpret the request as
the final XML request.

The fix is to simply use the latest URL to avoid the redirection.